slirp4netns: support using qemu for setting up the slirp network
it is not possible to use the system qemu for setting up and managing
the slirp network.

Add a --with-system-qemu build option, when it is specified the
libslirp code is not built in and the qemu executable is launched for
managing the network.  Some features like setting up the mtu are not
honored when system-qemu is used.

Closes: https://github.com/rootless-containers/slirp4netns/issues/30

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>