slirp4netns: add new option --create-sandbox
when running as root, either on the host or in a user namespace, add a
new option --create-sandbox that will create a new mount namespace and
bind mount only /etc and /run so that the network configuration can be
honored.  Everything else won't be accessible.  Also all capabilities
except CAP_NET_BIND_SERVICE are dropped.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>