seccompfilter: Conditional SCMP_ACT_KILL_PROCESS
As before, this only uses SCMP_ACT_KILL_PROCESS for libseccomp versions
which define it.
Additionally, it is only used if at run time the kernel reports it as a
supported action. For this SECCOMP_GET_ACTION_AVAIL must be available
(supported from Linux 4.14 onward and backported to CentOS 7's kernel)
and SECCOMP_RET_KILL_PROCESS (Linux 4.14+, but not in CentOS 7).

Signed-off-by: Marcel Bargull <marcel.bargull@udo.edu>