Support new --netns-type=tapfd
The new value tapfd enables removing slirp4netns features. Rather than
having it figure out where to create a tap network, the opened tap file
descriptor can now be inherited to slirp4netns and used as is. This
removes the need for forking and attaching a namespace as well as
readiness communication, but it also removes the ability to choose the
name of the network interface or configuring it.

Signed-off-by: Helmut Grohne <helmut@subdivi.de>